refactor: mitigate FOUC when applying Prism theme
