+++ packages/docusaurus-theme-classic/src/validateThemeConfig.ts
@@ -44,6 +44,7 @@ export const DEFAULT_CONFIG = {
   prism: {
     additionalLanguages: [],
     theme: defaultPrismTheme,
+    darkTheme: defaultPrismTheme,
     magicComments: [
       {
         className: 'theme-code-block-highlighted-line',
